+++ app/controllers/alchemy/admin/elements_controller.rb
@@ -90,14 +90,12 @@ module Alchemy
         end
       end
 
-      # Toggle fodls the element and persists the state in the db
+      # Toggle folds the element and persists the state in the db
       #
-      # Ingredient validations might make the element invalid.
-      # In this case we are just toggling a UI state and do not care about the validations.
       def fold
         @page = @element.page
-        @element.folded = !@element.folded
-        @element.save(validate: false)
+        # We do not want to trigger the touch callback or any validations
+        @element.update_columns(folded: !@element.folded)
       end
 
       private