+++ app/helpers/alchemy/url_helper.rb
@@ -20,8 +20,8 @@ module Alchemy
     def show_page_path_params(page, optional_params = {})
       raise ArgumentError, "Page is nil" if page.nil?
 
-      url_params = {urlname: page.urlname}.update(optional_params)
-      prefix_locale? ? url_params.update(locale: page.language_code) : url_params
+      url_params = { urlname: page.urlname }.update(optional_params)
+      prefix_locale?(page.language_code) ? url_params.update(locale: page.language_code) : url_params
     end
 
     # Returns the path for downloading an alchemy attachment
@@ -32,11 +32,6 @@ module Alchemy
     # Returns the url for downloading an alchemy attachment
     def download_alchemy_attachment_url(attachment)
       alchemy.download_attachment_url(attachment, attachment.slug)
-    end
-
-    # Returns the full url containing host, page and anchor for the given element
-    def full_url_for_element(element)
-      "#{current_server}/#{element.page.urlname}##{element_dom_id(element)}"
     end
   end
 end