+++ app/models/alchemy/essence_richtext.rb
@@ -13,6 +13,7 @@
 #
 
 module Alchemy
+  # @deprecated
   class EssenceRichtext < BaseRecord
     acts_as_essence preview_text_column: "stripped_body"